move snippets to tests
